requirements.txt
requests>=2.25.1
python-whois>=0.8.0
beautifulsoup4>=4.9.3
urllib3>=1.26.0
python-dateutil>=2.8.1
tldextract>=3.1.0